Update `chain_id` type and upgrade uniswap-sdk-core
The `chain_id` data type has been updated to `ChainId` in the Pool entity. Also, the `uniswap-sdk-core` dependency in Cargo.toml and Cargo.lock has been updated from version 0.10.0 to 0.11.0. These changes also required modifications in the `pool.rs` extension file.

+++ src/extensions/pool.rs
@@ -64,14 +64,14 @@ pub async fn get_pool<M: Middleware>(
     }
     Ok(Pool::new(
         token!(
-            chain_id as u32,
+            chain_id,
             token_a,
             token_a_decimals,
             token_a_symbol,
             token_a_name
         ),
         token!(
-            chain_id as u32,
+            chain_id,
             token_b,
             token_b_decimals,
             token_b_symbol,
